+++ changelog.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+### 0.0.5
+- new pom.xml configuration `useSwagger`: Include swagger annotations that use field and table descriptions (default):
+- new pom.xml configuration `useJakartaValidation`: Includes Jakarta Validation annotations that refer to fields and tables (default):
+- updated dependencies versions
 
 ### 0.0.4
 - ⚠️ License Change: Starting from version 0.0.4, this project is licensed under the Apache License 2.0.